Move ClusterAssertDeassertRiscsTest and ClusterReadWriteL1Test to separate header file
